requirements.txt
django 
djangorestframework 
requests 
scipy 
numpy 
pandas
geopy
python-dotenv